requirements.txt
fastapi
uvicorn
requests
python-dotenv
openai
pypdf
numpy
scikit-learn
supabase
